+++ test/workout_files_test.dart
@@ -1,0 +1,53 @@
+import 'dart:io';
+
+import 'package:csv/csv.dart';
+import 'package:fit_tool/fit_tool.dart';
+import 'package:fit_tool/src/fit_file.dart';
+import 'package:test/test.dart';
+
+void main() {
+  group('workout file tests', () {
+    setUp(() {});
+
+    test('Decode course file with developer fields', () async {
+      final file = File('./test/data/stagesLink_28832.fit');
+      final bytes = await file.readAsBytes();
+      final fitFile = FitFile.fromBytes(bytes);
+    });
+
+    test('Decode workout repeat greater than step file', () async {
+      final file = File('./test/data/trainerroad_744490.fit');
+      final bytes = await file.readAsBytes();
+      final fitFile = FitFile.fromBytes(bytes);
+
+      for (var record in fitFile.records) {
+        final message = record.message;
+
+        if (message is WorkoutStepMessage) {
+          // print(message.toRow());
+
+          switch (message.targetType) {
+            case WorkoutStepTarget.powerLap:
+              // print(
+              //     'low power: ${message.customTargetPowerLow}, low value: ${message.customTargetValueLow}');
+              // print(
+              //     'high power: ${message.customTargetPowerHigh}, high value: ${message.customTargetValueHigh}');
+              break;
+            default:
+          }
+
+          switch (message.durationType) {
+            case WorkoutStepDuration.time:
+              // print('duration time: ${message.durationTime}');
+              break;
+
+            case WorkoutStepDuration.open:
+              // print('duration open: ${message.durationValue}');
+              break;
+            default:
+          }
+        }
+      }
+    });
+  });
+}